(refactoring) Перенос MonthDayType в Enums.
- Так же инкримент минорной версии.
