+++ src/EconomyMonitor.Abstacts/IMonthDay.cs
@@ -1,4 +1,4 @@
-using EconomyMonitor.Primitives;
+using EconomyMonitor.Primitives.Enums;
 
 namespace EconomyMonitor.Abstacts;
