+++ src/EconomyMonitor.Primitives/version.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://raw.githubusercontent.com/dotnet/Nerdbank.GitVersioning/master/src/NerdBank.GitVersioning/version.schema.json",
-  "version": "0.0",
+  "version": "0.1",
   "assemblyVersion": {
     "precision": "revision"
   },
